% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_correlation.R
\name{analyze_correlation}
\alias{analyze_correlation}
\title{Analyse Data for}
\usage{
analyze_correlation(dat)
}
\arguments{
\item{n_sub}{A single numeric value specifying for how many subjects data should be simulated.}

\item{n_trials}{A single numeric value specifying how many trials per condition should be simulated.}

\item{par_limits}{A data.frame containing two variables: min & max with 7 values
for each parameter of the DMC. You have to either include a third variable `par_name` labeling
which row contains information for which parameters or name the rows of the data frame.
Parameter names of the dmc parameters can be obtained via: `dRiftDM::dmc_dm()$free_prms`}
}
\description{
Analyse Data for
}
