% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptives.R
\name{get_descriptives}
\alias{get_descriptives}
\title{Analyse Data for}
\usage{
get_descriptives(data, format = "long")
}
\arguments{
\item{data}{A single numeric value specifying for how many subjects data should be simulated.}

\item{format}{A single numeric value specifying how many trials per condition should be simulated.}
}
\description{
Analyse Data for
}
