% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ezDM_dmc.R
\name{get_ezDM_dmc}
\alias{get_ezDM_dmc}
\title{Analyse Data for}
\usage{
get_ezDM_dmc(data, format = "long")
}
\arguments{
\item{data}{A single numeric value specifying for how many subjects data should be simulated.}

\item{format}{A single numeric value specifying how many trials per condition should be simulated.}
}
\description{
Analyse Data for
}
