% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_correlated_vars.R
\name{simulate_correlated_vars}
\alias{simulate_correlated_vars}
\title{Simulate correlated variables}
\usage{
simulate_correlated_vars(
  values,
  correlation,
  mean = 0,
  sd = 1,
  lb = -Inf,
  ub = Inf
)
}
\arguments{
\item{values}{A numeric vector of values from which to simulate the correlated variable.}

\item{correlation}{A numeric value between -1 and 1 indicating the correlation between the simulated variable and the given variable.}

\item{mean}{A numeric value indicating the mean of the simulated variable.}

\item{sd}{A numeric value indicating the standard deviation of the simulated variable.}

\item{lb}{A numeric value indicating the lower bound of the simulated variable.}

\item{ub}{A numeric value indicating the upper bound of the simulated variable.}
}
\value{
A numeric vector of simulated correlated variables.
}
\description{
This function simulates a correlated variable from a given variable.
}
\examples{
values <- rnorm(100, 0, 1)
correlated_values <- simulate_correlated_vars(values, correlation = 0.5, mean = 10, sd = 2, lb = 4, ub = 16)
cor(values, correlated_values)

}
