% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_correlation_ssp.R
\name{simulate_correlation_ssp}
\alias{simulate_correlation_ssp}
\title{Simulate Correlated Data for two tasks from the DMC}
\usage{
simulate_correlation_ssp(
  n_sub,
  n_trials,
  correlation,
  correlated_par,
  par_limits = NULL,
  verbose = 0
)
}
\arguments{
\item{n_sub}{A single numeric value specifying for how many subjects data should be simulated.}

\item{n_trials}{A single numeric value specifying how many trials per condition should be simulated.}

\item{correlation}{A single numeric value specifying the correlation between the parameters of the two tasks.}

\item{correlated_par}{A character value specifying which parameters should be correlated.}

\item{par_limits}{A data.frame containing two variables: min & max with 7 values
for each parameter of the DMC. You have to either include a third variable `par_name` labeling
which row contains information for which parameters or name the rows of the data frame.
Parameter names of the dmc parameters can be obtained via: `dRiftDM::dmc_dm()$free_prms`}
}
\description{
Simulate Correlated Data for two tasks from the DMC
}
